requirements.txt
google-generativeai
pydub
PyYAML 